+++ NEWS.md
@@ -1,5 +1,7 @@
-# TreeTools 1.9.2.9002 (development) #
+# TreeTools 1.9.2.9003 (development) #
 
+- Add `TipTimedTree()` function to display trees where leaves are associated
+  with absolute ages.
 - Add `is.TreeNumber()` method.
 - Support zero-edge trees in `as.Splits()`.
 - Support empty constraints in `AddUnconstrained()`.